Move existing parsers into subdirectory
