+++ lib/money/parser/accounting.rb
